refactor(chat-store): extract message storage responsibility
